Provisioning: Consolidate repository and connection controller tests
